INFRA-1615-slack-approvals-prevent-self-approval (#7)
New flow as follow:

SlackApproval repository changes:

1. Developer Portal and Templates keep the same.
2. SlackRequest class keep same as before (just reformat code using
black).
3. SlackProvision class main changes:
a. Two new action_id added: “Not allowed” and “Reject Response”.
b. Not allowed action_id in case the flag prevent_self_approval=True
comes in the payload and the user requester same as user executing the
action. A popup opens in this case for the requester showing information
on not possible executing the action for that user.
c. When the user click on Reject a popup (modal view) opens asking for
rejection reason, when submitted the action change to “Reject Response”,
the submit send the answer to the same endpoint for SlackProvision which
in turns call the ProvisionClass call() method but this time with
“Reject Response” which in turn call rejected() method, send message
with rejected reason to the requester and approver thread (as reply) and
update the status for both channels.
d. The approved action_id keeps the same flow.